requirements.txt
bcrypt==4.0.1
mysql-connector==2.2.9
tomli==2.0.1
